+++ src/cadqueryhelper/shape/__init__.py
@@ -22,3 +22,4 @@ from .cornerJoin import corner_join
 from .pinwheel import make_circular_points, interweave_lists, pinwheel
 from .trapezoid import trapezoid
 from .backdrop import backdrop
+from .jersey_barrier import jersey_barrier