+++ changes.md
@@ -14,6 +14,8 @@
 * Modified the random function is used for calculating variable height with the irregular grid. The new functions supports better type safety for floats as opposed to implicit casting to integers but the generated outputs are now different from previous versions of the library.
 * Regenerated all of the examples.
 * Added backdrop shape, example, stl, and documentation
+* Yoinked Jersey Barrier code from cqindustry and added it to the shapes.
+* Added license blocks
 
 ## 0.2.1
 * Added shape.coffin